+++ src/allpress/db/models.py
@@ -1,5 +1,0 @@
-from hashlib import md5
-
-from allpress.db.io import Transactions
-
-